File reorganization, basic UI changes
