+++ package.json
@@ -12,7 +12,6 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "@radix-ui/react-accordion": "^1.2.3",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/themes": "^3.1.6",
     "cmdk": "^1.0.4",
