+++ src/pages/ProjectOverview/index.ts
@@ -1,3 +1,3 @@
-export * from "./Filters";
+export * from "./ProjectSearch";
 export * from "./ProjectCard";
 export * from "./Overview";
