+++ src/pages/VersionSelector/VersionCards.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex, Strong } from "@radix-ui/themes";
+import { Box, Flex, Strong, Link } from "@radix-ui/themes";
 import React from "react";
 import { Version } from "../../state/core";
 import { VersionCard } from "../ProjectOverview";
@@ -20,7 +20,8 @@ export const VersionCards: React.FC<VersionListProps> = ({
           version={version}
           title={
             <Box>
-              <Strong>Version: </Strong> {version.name}
+              <Strong>Version:</Strong>{" "}
+              <Link onClick={() => onVersionClick(index)}>{version.name}</Link>
             </Box>
           } // Only shows version name
           onClick={() => onVersionClick(index)}
